% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smry.R
\name{smry}
\alias{smry}
\title{Summary of numeric vectors}
\usage{
smry(x, y = 5)
}
\arguments{
\item{x}{a vector of numbers}

\item{y}{a second number}
}
\value{
A vector with the length (n), mean, and median.
}
\description{
Returns the length, mean, and median of a vectorof numbers
}
\examples{
# summarize 100 normally distributed numbers with mean of 5 and sd of 1
x.vec <- rnorm(100, 5, 1)
smry(x,vec)

\dontrun{
# can't use with characters
smry(letters)}


}
