% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{smrzvectors}
\alias{smrzvectors}
\title{Summarize Numeric vectors}
\usage{
smrzvectors(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
mean, median, and variance of vector
}
\description{
Gives summaries of a set of numbers
}
